0034774: Failed test: Vorschau/Bearbeitungsmodus einer Hotspot/Imagemap-Frage aufrufen

+++ Modules/TestQuestionPool/classes/class.assImagemapQuestion.php
@@ -979,9 +979,9 @@ class assImagemapQuestion extends assQuestion implements ilObjQuestionScoringAdj
 
     /**
      * @param $found_values
-     * @return int
+     * @return float
      */
-    protected function calculateReachedPointsForSolution($found_values): int
+    protected function calculateReachedPointsForSolution($found_values): float
     {
         $points = 0;
         if (count($found_values) > 0) {
