+++ Modules/TestQuestionPool/classes/class.assMultipleChoice.php
@@ -1282,9 +1282,9 @@ class assMultipleChoice extends assQuestion implements ilObjQuestionScoringAdjus
     /**
      * @param $found_values
      * @param $active_id
-     * @return int
+     * @return float
      */
-    protected function calculateReachedPointsForSolution($found_values, $active_id = 0): int
+    protected function calculateReachedPointsForSolution($found_values, $active_id = 0): float
     {
         $points = 0;
         foreach ($this->answers as $key => $answer) {
